+ fixed: - 0000244: [Bug] Runde Angabe der Genauigkeit + fixed: - 0000243: [Bug] Cachebox überprüft beim Start nicht, ob das GPS ausgeschaltet ist + fixed: - 0000240: [Bug] Beim Wechsel vom Tag/Modus wird das GPS (scheinbar?) neu initialisiert
git-svn-id: svn://svn.code.sf.net/p/droidcachebox/code/trunk@377 5890f383-b09f-485f-8f59-22bc967d67fd

+++ Android_GUI/src/de/droidcachebox/Views/AboutView.java
@@ -309,7 +309,7 @@ public class AboutView extends FrameLayout implements ViewOptionsMenu, SelectedC
 	{
 		 if ((Global.Locator.getLocation() != null) && (Global.Locator.getLocation().hasAccuracy()))
 	        {
-	        	float radius = Global.Locator.getLocation().getAccuracy();
+	        	int radius = (int) Global.Locator.getLocation().getAccuracy();
 	        	Accuracy.setText("+/- " + String.valueOf(radius) + "m (" + Global.Locator.ProviderString()+")");
 	        }
 		 else
